+++ src/core/compilers/svelte.ts
@@ -1,4 +1,3 @@
-import { escapeSvelte } from '../utils'
 import { Compiler } from './types'
 
 export const SvelteCompiler = <Compiler>((svg: string) => {
@@ -9,3 +8,12 @@ export const SvelteCompiler = <Compiler>((svg: string) => {
   const closeTag = svg.slice(closeTagStart)
   return `${openTag}${content}${closeTag}`
 })
+
+// escape curlies, backtick, \t, \r, \n to avoid breaking output of {@html `here`} in .svelte
+export function escapeSvelte(str: string): string {
+  return str
+    .replace(/{/g, '&#123;')
+    .replace(/}/g, '&#125;')
+    .replace(/`/g, '&#96;')
+    .replace(/\\([trn])/g, ' ')
+}
