+++ src/core/utils.ts
@@ -14,12 +14,3 @@ export function camelToKebab(key: string) {
     .trim()
   return result.split(/\s+/g).join('-').toLowerCase()
 }
-
-// escape curlies, backtick, \t, \r, \n to avoid breaking output of {@html `here`} in .svelte
-export function escapeSvelte(str: string): string {
-  return str
-    .replace(/{/g, '&#123;')
-    .replace(/}/g, '&#125;')
-    .replace(/`/g, '&#96;')
-    .replace(/\\([trn])/g, ' ')
-}
